review: address PR #48 second-round review findings
Verified each finding against the current code; fixed real bugs and
spec-vs-code mismatches, skipped one non-issue with justification.

Real bugs fixed:

- db::mediation::latest_open_session_for: mirrored the Feature 005
  carve-out from `list_live_sessions`. A `summary_delivered`
  session that received the cooperative invitation now also surfaces
  through the open-session lookup, so a re-take or a re-mediation
  attempt observes the same session-state semantics consistently
  across both queries. Legacy `summary_delivered` sessions stay
  terminal (no `self_resolution_offered` row).
- mediation::follow_up: when the post-invitation early-return arm
  fires (re-classified summary_delivered session whose decision is
  not `Escalate`), advance `round_count_last_evaluated` before
  returning. Without the advance the FR-127 idempotency gate would
  re-classify the same fresh inbound on every subsequent tick,
  burning reasoning-provider budget on a session that's already
  settled into "wait silently for a possible human-assistance
  request".
- mediation::policy: switch the cooperative threshold comparison to
  `confidence >= f64::from(self_resolution_threshold)` so the
  classifier's f64 confidence is no longer cast down to f32 before
  comparison. Threshold storage stays f32 per the config contract;
  the comparison is now precision-honest.

Spec-vs-code mismatches fixed (audit-row payload):

- db::mediation_events::record_self_resolution_offered: payload
  shape now matches `contracts/audit-events.md`:
  * `classification_confidence` (renamed from `confidence`)
  * `rationale_id` lives inside `payload_json` (the dedicated
    `mediation_events.rationale_id` column stays NULL on this kind
    — the column convention is reserved for rationales whose
    lifecycle is owned by the audit row itself; here the rationale
    is owned by the round-N classification call).
  * `languages` carries only `buyer` and `seller` codes (dropped
    `fallback`, which was code-internal noise).
  * Signature changed from `&Connection` to `&Transaction<'_>` so
    the audit row can only be written inside an outer transaction
    — same shape and discipline as `record_escalation_dispatched`.
    The caller in `follow_up.rs` already passes `&tx`, so this
    enforcement is purely additive.

Documentation fixes (verified against shipped code):

- config.sample.toml: clarified the `self_resolution_enabled = false`
  description. With the kill-switch off, Serbero falls back to the
  legacy cooperative-summary path (LLM still summarizes for the
  solver); it does NOT force every cooperative case through human
  review as the previous comment overstated.
- contracts/classifier-output.md: clarified the serde-defaults
  story. `ClassificationResponse` itself has no serde derives — it
  is the adapter-facing public shape, not a wire type. The actual
  wire-format deserialization lives in each adapter's intermediate
  struct (`ClassificationJson` in src/reasoning/openai.rs), which
  carries the `#[serde(default)]` annotations on `human_requested`
  / `buyer_language` / `seller_language`. Adapters are responsible
  for translating the wire struct into a `ClassificationResponse`.
- contracts/template-bundle.md: corrected the "ASCII-folded" claim.
  The keyword-audit comparison uses `str::to_ascii_lowercase` on
  both sides, which folds ASCII case but preserves diacritics
  (`ñ`, `ç`, `á`). The banned list enumerates diacritic-bearing
  forms verbatim — adding Unicode normalization would pull a new
  dependency for negligible coverage gain. The representative
  banned-list table in the spec now matches the canonical
  `BANNED` matrix in the test file.
- contracts/audit-events.md: relaxed the "Steps 3 and 4 land in the
  same transaction" claim to accurately describe the
  transactional-outbox behavior. Step 3 (`self_resolution_offered`)
  + the two outbound `mediation_messages` rows commit atomically;
  step 4 (`summary_generated`) lands in a subsequent transaction
  because the summarizer makes an LLM HTTP call between the two.
  A crash between TX1 and TX2 leaves the audit row + outbound rows
  in place, and the next tick's idempotency check sees the
  `self_resolution_offered` row + falls through to the legacy
  Summarize path to recover the still-pending summary.

Skipped (verified non-issue):

- The TOCTOU race re-flagged on the `prior_offered` predicate in
  `policy::evaluate` vs. the later write in `follow_up::dispatch`.
  In single-process operation the engine holds the
  `Arc<AsyncMutex<Connection>>` across the predicate read and the
  eventual write inside `draft_and_send_self_resolution_invitation`'s
  transaction, so two concurrent `advance_session_round` calls
  cannot both pass the predicate. Adding a UNIQUE partial index
  would require a new SQL migration which `plan.md` forbids; an
  HA / multi-process deploy is out of scope for this feature.

+++ src/mediation/policy.rs
@@ -319,7 +319,11 @@ pub async fn evaluate(
             confidence,
         } if mediation_cfg.self_resolution_enabled
             && templates_present
-            && (confidence as f32) >= mediation_cfg.self_resolution_threshold
+            // Compare in f64 to avoid losing precision on the
+            // higher-precision classifier confidence — the threshold
+            // is f32 by config-contract but the comparison stays
+            // honest.
+            && confidence >= f64::from(mediation_cfg.self_resolution_threshold)
             && !prior_offered =>
         {
             debug!(
